# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  httpErrors.RestError:
    properties:
      error:
        type: string
      status:
        type: integer
    type: object
...
  models.AccessToken:
    properties:
      prefixToken:
        type: string
      token:
        type: string
    type: object
...
  models.RefreshToken:
    properties:
      token:
        type: string
    type: object
  models.Tokens:
    properties:
      accessToken:
        $ref: '#/definitions/models.AccessToken'
      refreshToken:
        $ref: '#/definitions/models.RefreshToken'
    type: object
  models.User:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        type: string
      role:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    required:
    - name
    - password
    type: object
  models.UserBase:
    properties:
      user:
        $ref: '#/definitions/models.User'
    type: object
...
  models.UsersList:
    properties:
      has_more:
        type: boolean
      page:
        type: integer
      size:
        type: integer
      total_count:
        type: integer
      total_pages:
        type: integer
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API
  title: REST API
  version: "1.0"
paths:
...
  /auth/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: user_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestError'
      security:
      - Auth: []
      summary: Удалить учетную запись пользователя
      tags:
      - Auth
    get:
      consumes:
      - application/json
      parameters:
      - description: user_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestError'
      summary: Получить пользователя по идентификатору
      tags:
      - Auth
    put:
      consumes:
      - application/json
      description: Обновить существующего пользователя
      parameters:
      - description: user_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      security:
      - Auth: []
      summary: Обновить пользователя
      tags:
      - Auth
  /auth/{id}/avatar:
    post:
      consumes:
      - application/json
      parameters:
      - description: Body with image file
        in: formData
        name: file
        required: true
        type: file
      - description: aws s3 bucket
        format: bucket
        in: query
        name: bucket
        required: true
        type: string
      - description: user_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestError'
      security:
      - Auth: []
      summary: Опубликовать аватар пользователя
      tags:
      - Auth
  /auth/all:
    get:
      consumes:
      - application/json
      parameters:
      - description: page number
        format: page
        in: query
        name: page
        type: integer
      - description: number of elements per page
        format: size
        in: query
        name: size
        type: integer
      - description: filter name
        format: orderBy
        in: query
        name: orderBy
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UsersList'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestError'
      summary: Получить список всех пользователей
      tags:
      - Auth
  /auth/check:
    get:
      consumes:
      - application/json
      description: Проверить токен на действительность
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserBase'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestError'
      security:
      - Auth: []
      summary: Проверить зарегистрированного пользователя
      tags:
      - Auth
  /auth/confirmed:
    get:
      consumes:
      - application/json
      description: Подтверждается почта
      parameters:
      - description: code
        format: code
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestError'
      summary: Подтверждение почты
      tags:
      - Auth
  /auth/find:
    get:
      consumes:
      - application/json
      parameters:
      - description: username
        format: username
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UsersList'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestError'
      summary: Найти пользователя по имени
      tags:
      - Auth
...
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Обновить и вернуть новые токены
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Tokens'
      security:
      - Auth: []
      summary: Обновить токен
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Зарегистрировать и вернуть пользователя и токен
      parameters:
      - description: register user
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "201":
          description: ""
      summary: Зарегистрировать нового пользователя
      tags:
      - Auth
  /auth/token:
    get:
      consumes:
      - application/json
      description: Получить токен CSRF, обязательный cookie сеанса авторизации
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpErrors.RestError'
      security:
      - Auth: []
      summary: Получить токен CSRF
      tags:
      - Auth
...
securityDefinitions:
  Auth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"